+++ src/log.c
@@ -51,7 +51,7 @@ void scas_abort(char *format, ...) {
 
 void scas_log(int verbosity, char* format, ...) {
 	if (verbosity <= v) {
-		int c = verbosity;
+		size_t c = verbosity;
 		if (c > sizeof(verbosity_colors) / sizeof(char *)) {
 			c = sizeof(verbosity_colors) / sizeof(char *) - 1;
 		}
